3Dmaps made much faster and integration with settings solved
1. Now we have multiple datasets, and only the dataset which needs to be
rebuilt is done.  This makes selecting a star, changing grid nearly
immediate.
2. Cleaned up how its run.  Only two direct accesses left to the class.
no access to settings db items in there - all data is passed in
3. Fixed problem with route planner - route would not be updated if map
was already open
4.Slowed down movement a tad and also limited zoom factor in movement so
you can't get to a zero crawl speed.

there may be bugs. large change

+++ EDDiscovery/Settings.cs
@@ -20,7 +20,7 @@ namespace EDDiscovery2
         private ThemeEditor themeeditor = null;
 
         public string MapHomeSystem { get { return textBoxHomeSystem.Text; } }
-        public double MapZoom { get { return Double.Parse(textBoxDefaultZoom.Text); } }
+        public float MapZoom { get { return float.Parse(textBoxDefaultZoom.Text); } }
         public bool MapCentreOnSelection { get { return radioButtonHistorySelection.Checked; } }
 
         public Settings()
